move BaseMiner to its own file and abstract get_miner to pyasic.miners
