+++ pyasic/miners/_backends/btminer.py
@@ -18,7 +18,7 @@ from typing import Union
 
 
 from pyasic.API.btminer import BTMinerAPI
-from pyasic.miners import BaseMiner
+from pyasic.miners.base import BaseMiner
 from pyasic.API import APIError
 
 from pyasic.data import MinerData
