+++ pyasic/miners/_types/antminer/X17/S17_Pro.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners import BaseMiner
+from pyasic.miners.base import BaseMiner
 
 
 class S17Pro(BaseMiner):
