+++ pyasic/miners/_types/antminer/X19/S19_Pro.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners import BaseMiner
+from pyasic.miners.base import BaseMiner
 
 
 class S19Pro(BaseMiner):
