+++ pyasic/miners/_types/avalonminer/A8X/A821.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners import BaseMiner
+from pyasic.miners.base import BaseMiner
 
 
 class Avalon821(BaseMiner):
